shipyard_blueprints/motherduck/pyproject.toml
[tool.poetry]
name = "shipyard-motherduck"
version = "0.1.0a0"
description = "A local client for working with Python and MotherDuck"
authors = ["wrp801 <wespoulsen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
shipyard-templates = "^0.8.2"
shipyard-bp-utils = "^1.2.1"
duckdb = "0.9.2"
pandas = "^2.0"
sqlalchemy = "^2.0.29"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.1.1"
black = "^24.3.0"
pyarrow = "^15.0.2"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
